ahc.py
import requests
from bs4 import BeautifulSoup

# Replace with a Google Search URL for restaurants (avoid directly scraping Google Maps)
url = "https://www.google.com/search?q=restaurants+in+new+york"

# Fetch the webpage content
response = requests.get(url)
soup = BeautifulSoup(response.content, "html.parser")

# This is a simplified example, actual implementation might require more complex parsing 
reviews = []
for review in soup.find_all("div", class_="review-snippet"):
  text = review.find("span", class_="review-text").text.strip()
  rating = review.find("span", class_="business-rating").get("aria-label")
  reviews.append({"text": text, "rating": rating})

print(reviews)