requirements.txt
deepgram_sdk
langchain
langchain_community
langchain_core
langchain_groq
langchain_openai
python-dotenv
Requests
streamlit
openai 
jq
chromadb
pysqlite3-binary
langsmith
python-dotenv
pypdf
faiss-cpu
google-generativeai
fastapi
uvicorn
twilio
python-decouple
sqlalchemy
psycopg2-binary
python-multipart
pyngrok
flask